test: Add nav section tests
